// src/generated/wotlk/api.ts
export interface Party {
    /**
     * @generated from protobuf field: repeated proto.Player players = 1
     */
    players: Player[];
    /**
     * @generated from protobuf field: proto.PartyBuffs buffs = 2
     */
    buffs?: PartyBuffs;
}
/**
 * @generated from protobuf message proto.Raid
 */
export interface Raid {
    /**
     * @generated from protobuf field: repeated proto.Party parties = 1
     */
    parties: Party[];
    /**
     * The number of parties in the raid who participate in the sim, e.g.
     * 10-man vs 25-man raids. The remaining parties are unused but can hold
     * bench players.
     *
     * If this value is 0, all parties are included.
     *
     * @generated from protobuf field: int32 num_active_parties = 2
     */
    numActiveParties: number;
    /**
     * @generated from protobuf field: proto.RaidBuffs buffs = 7
     */
    buffs?: RaidBuffs;
    /**
     * Extra debuffs provided by buff bots in this raid.
     * This is also used for debuffs in the individual sims.
     *
     * @generated from protobuf field: proto.Debuffs debuffs = 5
     */
    debuffs?: Debuffs;
    /**
     * Players who will be tanking mobs.
     *
     * @generated from protobuf field: repeated proto.UnitReference tanks = 4
     */
    tanks: UnitReference[];
    /**
     * Staggers Stormstrike casts across Enhance Shaman to maximize charge usage.
     *
     * @generated from protobuf field: bool stagger_stormstrikes = 3
     */
    staggerStormstrikes: boolean;
    /**
     * Extra fake players to add. Currently only used by healing sims.
     *
     * @generated from protobuf field: int32 target_dummies = 6
     */
    targetDummies: number;
}
/**
 * @generated from protobuf message proto.SimOptions
 */
export interface SimOptions {
    /**
     * @generated from protobuf field: int32 iterations = 1
     */
    iterations: number;
    /**
     * @generated from protobuf field: int64 random_seed = 2
     */
    randomSeed: bigint;
    /**
     * @generated from protobuf field: bool debug = 3
     */
    debug: boolean; // Enables debug logging.
    /**
     * @generated from protobuf field: bool debug_first_iteration = 6
     */
    debugFirstIteration: boolean;
    /**
     * @generated from protobuf field: bool is_test = 5
     */
    isTest: boolean; // Only used internally.
    /**
     * @generated from protobuf field: bool save_all_values = 7
     */
    saveAllValues: boolean; // Only used internally.
    /**
     * @generated from protobuf field: bool interactive = 8
     */
    interactive: boolean; // Enables interactive mode.
}
// ...
export interface PartyMetrics {
    /**
     * @generated from protobuf field: proto.DistributionMetrics dps = 1
     */
    dps?: DistributionMetrics;
    /**
     * @generated from protobuf field: proto.DistributionMetrics hps = 3
     */
    hps?: DistributionMetrics;
    /**
     * @generated from protobuf field: repeated proto.UnitMetrics players = 2
     */
    players: UnitMetrics[];
}
/**
 * Results for a whole raid.
 *
 * @generated from protobuf message proto.RaidMetrics
 */
export interface RaidMetrics {
    /**
     * @generated from protobuf field: proto.DistributionMetrics dps = 1
     */
    dps?: DistributionMetrics;
    /**
     * @generated from protobuf field: proto.DistributionMetrics hps = 3
     */
    hps?: DistributionMetrics;
    /**
     * @generated from protobuf field: repeated proto.PartyMetrics parties = 2
     */
    parties: PartyMetrics[];
}
/**
 * @generated from protobuf message proto.EncounterMetrics
 */
export interface EncounterMetrics {
    /**
     * @generated from protobuf field: repeated proto.UnitMetrics targets = 1
     */
    targets: UnitMetrics[];
}
/**
 * RPC RaidSim
 *
 * @generated from protobuf message proto.RaidSimRequest
 */
export interface RaidSimRequest {
    /**
     * @generated from protobuf field: proto.Raid raid = 1
     */
    raid?: Raid;
    /**
     * @generated from protobuf field: proto.Encounter encounter = 2
     */
    encounter?: Encounter;
    /**
     * @generated from protobuf field: proto.SimOptions sim_options = 3
     */
    simOptions?: SimOptions;
}
/**
 * Result from running the raid sim.
 *
 * @generated from protobuf message proto.RaidSimResult
 */
export interface RaidSimResult {
    /**
     * @generated from protobuf field: proto.RaidMetrics raid_metrics = 1
     */
    raidMetrics?: RaidMetrics;
    /**
     * @generated from protobuf field: proto.EncounterMetrics encounter_metrics = 2
     */
    encounterMetrics?: EncounterMetrics;
    /**
     * @generated from protobuf field: string logs = 3
     */
    logs: string;
    /**
     * Needed for displaying the timeline properly when the duration +/- option
     * is used.
     *
     * @generated from protobuf field: double first_iteration_duration = 4
     */
    firstIterationDuration: number;
    /**
     * @generated from protobuf field: double avg_iteration_duration = 6
     */
    avgIterationDuration: number;
    /**
     * @generated from protobuf field: string error_result = 5
     */
    errorResult: string;
}
// ...
export interface RaidStats {
    /**
     * @generated from protobuf field: repeated proto.PartyStats parties = 1
     */
    parties: PartyStats[];
}
/**
 * @generated from protobuf message proto.TargetStats
 */
export interface TargetStats {
    /**
     * @generated from protobuf field: proto.UnitMetadata metadata = 1
     */
    metadata?: UnitMetadata;
}
/**
 * @generated from protobuf message proto.EncounterStats
 */
export interface EncounterStats {
    /**
     * @generated from protobuf field: repeated proto.TargetStats targets = 1
     */
    targets: TargetStats[];
}
/**
 * @generated from protobuf message proto.ComputeStatsResult
 */
export interface ComputeStatsResult {
    /**
     * @generated from protobuf field: proto.RaidStats raid_stats = 1
     */
    raidStats?: RaidStats;
    /**
     * @generated from protobuf field: proto.EncounterStats encounter_stats = 3
     */
    encounterStats?: EncounterStats;
    /**
     * @generated from protobuf field: string error_result = 2
     */
    errorResult: string;
}
// ...
export const Party = new Party$Type();
// ...
export const Raid = new Raid$Type();
// ...
export const SimOptions = new SimOptions$Type();
// ...
export const DistributionMetrics = new DistributionMetrics$Type();
// ...
export const PartyMetrics = new PartyMetrics$Type();
// ...
export const RaidMetrics = new RaidMetrics$Type();
// ...
export const EncounterMetrics = new EncounterMetrics$Type();
// @generated message type with reflection information, may provide speed optimized methods
class RaidSimRequest$Type extends MessageType<RaidSimRequest> {
    constructor() {
        super("proto.RaidSimRequest", [
            { no: 1, name: "raid", kind: "message", T: () => Raid },
            { no: 2, name: "encounter", kind: "message", T: () => Encounter },
            { no: 3, name: "sim_options", kind: "message", T: () => SimOptions }
        ]);
    }
    create(value?: PartialMessage<RaidSimRequest>): RaidSimRequest {
        const message = globalThis.Object.create((this.messagePrototype!));
        if (value !== undefined)
            reflectionMergePartial<RaidSimRequest>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: RaidSimRequest): RaidSimRequest {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* proto.Raid raid */ 1:
                    message.raid = Raid.internalBinaryRead(reader, reader.uint32(), options, message.raid);
                    break;
                case /* proto.Encounter encounter */ 2:
                    message.encounter = Encounter.internalBinaryRead(reader, reader.uint32(), options, message.encounter);
                    break;
                case /* proto.SimOptions sim_options */ 3:
                    message.simOptions = SimOptions.internalBinaryRead(reader, reader.uint32(), options, message.simOptions);
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: RaidSimRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* proto.Raid raid = 1; */
        if (message.raid)
            Raid.internalBinaryWrite(message.raid, writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        /* proto.Encounter encounter = 2; */
        if (message.encounter)
            Encounter.internalBinaryWrite(message.encounter, writer.tag(2, WireType.LengthDelimited).fork(), options).join();
        /* proto.SimOptions sim_options = 3; */
        if (message.simOptions)
            SimOptions.internalBinaryWrite(message.simOptions, writer.tag(3, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message proto.RaidSimRequest
 */
export const RaidSimRequest = new RaidSimRequest$Type();
// ...
export const RaidSimResult = new RaidSimResult$Type();
// ...
export const RaidStats = new RaidStats$Type();
// ...
export const EncounterStats = new EncounterStats$Type();
// @generated message type with reflection information, may provide speed optimized methods
class ComputeStatsResult$Type extends MessageType<ComputeStatsResult> {
    constructor() {
        super("proto.ComputeStatsResult", [
            { no: 1, name: "raid_stats", kind: "message", T: () => RaidStats },
            { no: 3, name: "encounter_stats", kind: "message", T: () => EncounterStats },
            { no: 2, name: "error_result", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
        ]);
    }
    create(value?: PartialMessage<ComputeStatsResult>): ComputeStatsResult {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.errorResult = "";
        if (value !== undefined)
            reflectionMergePartial<ComputeStatsResult>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: ComputeStatsResult): ComputeStatsResult {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* proto.RaidStats raid_stats */ 1:
                    message.raidStats = RaidStats.internalBinaryRead(reader, reader.uint32(), options, message.raidStats);
                    break;
                case /* proto.EncounterStats encounter_stats */ 3:
                    message.encounterStats = EncounterStats.internalBinaryRead(reader, reader.uint32(), options, message.encounterStats);
                    break;
                case /* string error_result */ 2:
                    message.errorResult = reader.string();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: ComputeStatsResult, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* proto.RaidStats raid_stats = 1; */
        if (message.raidStats)
            RaidStats.internalBinaryWrite(message.raidStats, writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        /* string error_result = 2; */
        if (message.errorResult !== "")
            writer.tag(2, WireType.LengthDelimited).string(message.errorResult);
        /* proto.EncounterStats encounter_stats = 3; */
        if (message.encounterStats)
            EncounterStats.internalBinaryWrite(message.encounterStats, writer.tag(3, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message proto.ComputeStatsResult
 */
export const ComputeStatsResult = new ComputeStatsResult$Type();